Working version - age is now short instead of byte (because 127 is too low in rare cases)

+++ src/at/sume/dm/model/output/OutputPersonRow.java
@@ -14,7 +14,7 @@ public class OutputPersonRow implements OutputRow {
 	private int personId;
 	private int householdId;
 	private byte sex;
-	private byte age;
+	private short age;
 	private int yearlyIncome;
 
 	public OutputPersonRow(short modelYear, PersonRow person) {
